Using slot
