+++ packages/@orchidui-vue/src/Elements/Header/OcHeader.stories.js
@@ -7,6 +7,9 @@ import {
   HeaderLeft,
   HeaderRight,
   Icon,
+  Dropdown,
+  DropdownItem,
+  Button,
 } from "@/orchidui";
 
 import SampleHeaderLeft from "./SampleHeaderLeft.vue";
@@ -146,6 +149,9 @@ export const SubHeaderDropdownElement = {
       HeaderCenter,
       HeaderRight,
       SampleHeaderLeft,
+      Dropdown,
+      DropdownItem,
+      Button,
     },
     setup() {
       const activeMenuValue = ref("payments");
@@ -153,7 +159,17 @@ export const SubHeaderDropdownElement = {
         return args.menus.find((m) => m.value === activeMenuValue.value).label;
       });
 
-      return { args, activeMenuValue, activeMenuLabel };
+      const isDropdownOpen = ref(false);
+
+      const onDropdownItemClick = (action) => {
+        if (!action) return
+
+        console.log('dropdown-action', action)
+        isDropdownOpen.value = false
+      }
+
+
+      return { args, activeMenuValue, activeMenuLabel, isDropdownOpen, onDropdownItemClick };
     },
     template: `
           <Theme>
@@ -162,7 +178,34 @@ export const SubHeaderDropdownElement = {
                 <SampleHeaderLeft is-sub-header/>
               </HeaderLeft>
               <HeaderCenter class="flex-1" :is-saved="args.isSaved" />
-              <HeaderRight :is-saved="args.isSaved" :primary-props="args.primaryProps" :dropdown-options="args.dropdownOptions" />
+              <HeaderRight :is-saved="args.isSaved" :primary-props="args.primaryProps">
+                <template #primary-button>
+                  <Dropdown v-model="isDropdownOpen" placement="bottom-end">
+                    <Button
+                      class="min-w-[100px]"
+                      label="Save"
+                      v-bind="args.primaryProps"
+                      is-additional-area
+                      additional-area-icon="chevron-down"
+                      @click="$emit('save')"
+                      @addition-click="isDropdownOpen = true"
+                    />
+                    <template #menu>
+                      <div class="flex flex-col">
+                        <div class="p-2 border-b border-gray-200">
+                          <DropdownItem
+                            v-for="option, index in args.dropdownOptions"
+                            :key="index"
+                            :text="option.label"
+                            :icon="option.icon ?? null"
+                            @click="onDropdownItemClick(option.action)"
+                          />
+                        </div>
+                      </div>
+                    </template>
+                  </Dropdown>
+                </template>
+              </HeaderRight>
             </SubHeader>
           </Theme>
         `,
